Add updateCart method, get cart items method change
